+++ R/methods_ontology.R
@@ -1,19 +1,69 @@
-# methods ----
+# methods ----------------------------------------------------------------------
 
-#' Calculate the Resnik and Lin semantic similarity for an ontology.
+## pre-processing --------------------------------------------------------------
+
+#' Pre-process data for subsequent ontology similarity
+#'
+#' @description This function calculates needed information for semantic
+#' similiary calculations
+#'
+#' @param object `ontology class`. See [bixverse::ontology()].
+#' @param .verbose Boolean. Controls verbosity of the function.
+#'
+#' @return The class with added pre-processed data for semantic similarities to
+#' the properties.
+#'
+#' @export
+pre_process_sim_onto <- S7::new_generic(
+  name = "pre_process_sim_onto",
+  dispatch_args = "object",
+  fun = function(object, .verbose = TRUE) {
+    S7::S7_dispatch()
+  }
+)
+
+#' @export
+#'
+#' @import data.table
+#' @importFrom magrittr `%>%`
+#'
+#' @method pre_process_sim_onto ontology
+S7::method(pre_process_sim_onto, ontology) <- function(
+  object,
+  .verbose = TRUE
+) {
+  # Checks
+  checkmate::assertClass(object, "bixverse::ontology")
+  checkmate::qassert(.verbose, "B1")
+
+  parent_child_dt <- S7::prop(object, "parent_child_dt")
+
+  if (.verbose) {
+    message("Identifying the ancestors in the ontology.")
+  }
+  c(ancestors, descendants) %<-% get_ontology_ancestry(parent_child_dt)
+  if (.verbose) {
+    message("Calculating the information content of each term")
+  }
+  information_content <- calculate_information_content(descendants)
+
+  S7::prop(object, "outputs")[['ancestors']] <- ancestors
+  S7::prop(object, "outputs")[['descendants']] <- descendants
+  S7::prop(object, "outputs")[['information_content']] <- information_content
+
+  return(object)
+}
+
+## similarities ----------------------------------------------------------------
+
+#' Calculate the Resnik or Lin semantic similarity for an ontology.
 #'
 #' @description This function calculates the specified semantic similarities for
-#' the whole ontology, calculate a critical values based on `random_sample_no`
-#' (based on a user-specified alpha) and return all the term-term pairs above
-#' that threshold.
+#' the whole ontology and adds it to the class.
 #'
 #' @param object `ontology class`. See [bixverse::ontology()].
 #' @param sim_type String. One of `c("resnik", "lin", "combined")`.
-#' @param alpha Float. Between 0 and 1. A value of 0.01 indicates that the
-#' critical value is at the 1% quantile of the random sample of similarities.
-#' @param random_sample_no Integer. Number of random samples to use to estimate
-#' the critical value.
-#' @param seed Integer. Random seed for sampling reproducibility.
+#' @param .verbose Boolean. Controls the verbosity of the function.
 #'
 #' @return The class with added semantic similarities to the properties.
 #'
@@ -23,10 +73,8 @@ calculate_semantic_sim_onto <- S7::new_generic(
   dispatch_args = "object",
   fun = function(
     object,
-    sim_type,
-    alpha = 0.01,
-    random_sample_no = 100000L,
-    seed = 42L
+    sim_type = c("resnik", "lin", "combined"),
+    .verbose = TRUE
   ) {
     S7::S7_dispatch()
   }
@@ -41,47 +89,50 @@ calculate_semantic_sim_onto <- S7::new_generic(
 S7::method(calculate_semantic_sim_onto, ontology) <-
   function(
     object,
-    sim_type,
-    alpha = 0.01,
-    random_sample_no = 100000L,
-    seed = 42L
+    sim_type = c("resnik", "lin", "combined"),
+    .verbose = TRUE
   ) {
+    sim_type <- match.arg(sim_type)
+
     # Checks
     checkmate::assertClass(object, "bixverse::ontology")
     checkmate::assertChoice(sim_type, c("resnik", "lin", "combined"))
-    checkmate::qassert(alpha, "R1[0,1]")
-    checkmate::qassert(random_sample_no, "I1")
-    checkmate::qassert(seed, "I1")
+    checkmate::qassert(.verbose, "B1")
 
-    ancestor_list <- S7::prop(object, "ancestor_list")
-    information_content_list <- S7::prop(object, "information_content_list")
+    if (is.null(S7::prop(object, "outputs")[['ancestors']])) {
+      warning(
+        paste(
+          "No pre-processed data for semantic similarity calculation found.",
+          "Running the function now."
+        )
+      )
+      object <- pre_process_sim_onto(object = object, .verbose = .verbose)
+    }
+
+    ancestor_list <- S7::prop(object, "outputs")[['ancestors']]
+    information_content_list <- S7::prop(object, "outputs")[[
+      'information_content'
+    ]]
     terms <- names(ancestor_list)
 
-    similarities <- rs_onto_similarity_filtered(
+    similarities <- rs_onto_semantic_sim(
       terms = terms,
       sim_type = sim_type,
-      alpha = alpha,
       ancestor_list = ancestor_list,
-      ic_list = information_content_list,
-      iters = random_sample_no,
-      seed = seed
+      ic_list = information_content_list
+    )
+
+    final_sim <- upper_triangular_cor_mat$new(
+      cor_coef = similarities,
+      shift = 1L,
+      features = terms
     )
 
     params <- list(
-      critval = similarities$critval,
-      random_sample_no = random_sample_no,
-      alpha = alpha,
-      seed = seed,
       sim_type = sim_type
     )
 
-    similarities_dt <- data.table::setDT(similarities[c(
-      "term1",
-      "term2",
-      "filtered_sim"
-    )])
-
-    S7::prop(object, "semantic_similarities") <- similarities_dt
+    S7::prop(object, "sim_mat") <- final_sim
     S7::prop(object, "params")[["semantic_similarity"]] <- params
 
     return(object)
@@ -91,10 +142,11 @@ S7::method(calculate_semantic_sim_onto, ontology) <-
 
 ## main functions --------------------------------------------------------------
 
-#' Calculate the Resnik and Lin semantic similarity
+#' Calculate the Resnik or Lin semantic similarity
 #'
 #' @description This function calculates the semantic similarities based on
-#' Resnik and Lin similarity for a given ontology.
+#' Resnik or Lin similarity for a given ontology. Has also the option to
+#' calculate a combined version of the two.
 #'
 #' @param terms Vector of strings. The terms in the ontology you wish to screen.
 #' @param similarity_type String. One of `c("resnik", "lin", "combined")`.
@@ -124,7 +176,7 @@ calculate_semantic_sim <- function(
   checkmate::assertList(ic_list, types = "double")
   checkmate::assertNames(names(ic_list), must.include = terms)
 
-  onto_similarities <- rs_onto_similarity(
+  onto_similarities <- rs_onto_semantic_sim(
     terms = terms,
     sim_type = similarity_type,
     ancestor_list = ancestor_list,
@@ -141,6 +193,41 @@ calculate_semantic_sim <- function(
   colnames(matrix) <- rownames(matrix) <- terms
 
   return(matrix)
+}
+
+
+#' Calculate the Wang similarity matrix
+#'
+#' @description This function calculates the Wang similarity, based on the DAG
+#' for a given ontology. The current implementation just allows for a single w
+#' for the relationships.
+#'
+#' @param parent_child_dt data.table. The data.table with column parent and
+#' child.
+#' @param w Float. The value assigned to the respective edge. Defaults to `0.8`.
+#'
+#' @return The symmetric Wang similarity matrix.
+#'
+#' @export
+#'
+#' @import data.table
+calculate_wang_sim <- function(parent_child_dt, w) {
+  # Checks
+  checkmate::assertDataTable(parent_child_dt)
+  checkmate::assert(all(c("parent", "child") %in% colnames(parent_child_dt)))
+  checkmate::qassert(w, "R1(0,1)")
+
+  sim_wang <- rs_onto_sim_wang(
+    parent_child_dt$parent,
+    parent_child_dt$child,
+    w = w
+  )
+
+  sim_wang_mat <- sim_wang$sym_mat %>%
+    `colnames<-`(sim_wang$names) %>%
+    `rownames<-`(sim_wang$names)
+
+  return(sim_wang_mat)
 }
 
 ## helpers ---------------------------------------------------------------------
@@ -173,11 +260,11 @@ get_ontology_ancestry <- function(parent_child_dt) {
   graph <- igraph::graph_from_data_frame(edge_df[, c("from", "to")])
   ancestor_DT <- graph %>%
     igraph::ego(order = igraph::vcount(graph), mode = "out") %>%
-    setNames(igraph::V(graph)$name) %>%
+    stats::setNames(igraph::V(graph)$name) %>%
     Map(f = names) %>%
     stack() %>%
     rev() %>%
-    setNames(c("from", "to")) %>%
+    stats::setNames(c("from", "to")) %>%
     data.table::as.data.table() %>%
     .[, lapply(.SD, as.character)]
 
@@ -211,7 +298,7 @@ calculate_information_content <- function(ancestor_list) {
   checkmate::assertNamed(ancestor_list)
 
   tab <- purrr::map_dbl(ancestor_list, length)
-  information_content <- setNames(
+  information_content <- stats::setNames(
     -log(as.integer(tab) / length(ancestor_list)),
     nm = names(ancestor_list)
   )